removed category model and made it an array insidel object model
