+++ server/routes/products.routes.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const { createProduct } = require("../controllers/products.controllers.js");
+
+const router = express.Router();
+
+router.post("/", createProduct);
+
+module.exports = router;
